@@ -16,31 +16,32 @@
   "author": "Gabriel Ruttner",
   "license": "ISC",
   "dependencies": {
-    "@types/bcrypt": "^3.0.0",
-    "@types/jsonwebtoken": "^8.3.2",
-    "aws-sdk": "^2.492.0",
-    "bcrypt": "^5.0.0",
-    "body-parser": "^1.19.0",
+    "aws-sdk": "^2.1194.0",
+    "bcrypt": "^5.0.1",
+    "body-parser": "^1.20.0",
+    "dotenv": "^16.0.1",
     "email-validator": "^2.0.4",
-    "express": "^4.17.1",
+    "express": "^4.18.1",
     "jsonwebtoken": "^8.5.1",
-    "pg": "^7.11.0",
+    "pg": "^8.7.3",
     "reflect-metadata": "^0.1.13",
-    "sequelize": "^5.10.0",
-    "sequelize-typescript": "^0.6.11"
+    "sequelize": "^6.21.3",
+    "sequelize-typescript": "^2.1.3"
   },
   "devDependencies": {
-    "@types/bluebird": "^3.5.33",
-    "@types/express": "^4.17.9",
-    "@types/node": "^11.15.42",
-    "@types/sequelize": "^4.28.9",
-    "@types/validator": "^10.11.3",
-    "chai": "^4.2.0",
+    "@types/bcrypt": "^5.0.0",
+    "@types/bluebird": "^3.5.36",
+    "@types/express": "^4.17.13",
+    "@types/jsonwebtoken": "^8.5.8",
+    "@types/node": "^18.7.3",
+    "@types/sequelize": "^4.28.14",
+    "@types/validator": "^13.7.5",
+    "chai": "^4.3.6",
     "chai-http": "^4.3.0",
-    "mocha": "^6.2.3",
+    "mocha": "^10.0.0",
     "rimraf": "^3.0.2",
-    "ts-node-dev": "^1.1.1",
-    "tslint": "^5.20.1",
-    "typescript": "^3.9.7"
+    "ts-node-dev": "^2.0.0",
+    "tslint": "^6.1.3",
+    "typescript": "^4.7.4"
   }
 }
