@@ -1,20 +1,21 @@
-import AWS = require('aws-sdk');
-import { config } from './config/config';
+import AWS = require("aws-sdk");
+import { config } from "./config/config";
 
 const c = config.dev;
 
 //Configure AWS
-if(c.aws_profile !== "DEPLOYED") {
-  var credentials = new AWS.SharedIniFileCredentials({profile: c.aws_profile});
- AWS.config.credentials = credentials;
+if (c.aws_profile !== "DEPLOYED") {
+  var credentials = new AWS.SharedIniFileCredentials({
+    profile: c.aws_profile,
+  });
+  AWS.config.credentials = credentials;
 }
 
 export const s3 = new AWS.S3({
-  signatureVersion: 'v4',
+  signatureVersion: "v4",
   region: c.aws_region,
-  params: {Bucket: c.aws_media_bucket}
+  params: { Bucket: c.aws_media_bucket },
 });
-
 
 /* getGetSignedUrl generates an aws signed url to retreive an item
  * @Params
@@ -22,17 +23,16 @@ export const s3 = new AWS.S3({
  * @Returns:
  *    a url as a string
  */
-export function getGetSignedUrl( key: string ): string{
+export function getGetSignedUrl(key: string): string {
+  const signedUrlExpireSeconds = 60 * 5;
 
-  const signedUrlExpireSeconds = 60 * 5
+  const url = s3.getSignedUrl("getObject", {
+    Bucket: c.aws_media_bucket,
+    Key: key,
+    Expires: signedUrlExpireSeconds,
+  });
 
-    const url = s3.getSignedUrl('getObject', {
-        Bucket: c.aws_media_bucket,
-        Key: key,
-        Expires: signedUrlExpireSeconds
-      });
-
-    return url;
+  return url;
 }
 
 /* getPutSignedUrl generates an aws signed url to put an item
@@ -41,15 +41,21 @@ export function getGetSignedUrl( key: string ): string{
  * @Returns:
  *    a url as a string
  */
-export function getPutSignedUrl( key: string ){
+export function getPutSignedUrl(key: string) {
+  const signedUrlExpireSeconds = 60 * 5;
 
-    const signedUrlExpireSeconds = 60 * 5
+  console.log("REGION: ", c.aws_region);
+  console.log("KEY: ", key);
+  console.log("AWS: ", c.aws_profile);
+  console.log("S3: ", c.aws_media_bucket);
 
-    const url = s3.getSignedUrl('putObject', {
-      Bucket: c.aws_media_bucket,
-      Key: key,
-      Expires: signedUrlExpireSeconds
-    });
+  const url = s3.getSignedUrl("putObject", {
+    Bucket: c.aws_media_bucket,
+    Key: key,
+    Expires: signedUrlExpireSeconds,
+  });
 
-    return url;
+  console.log("PUT: ", url);
+
+  return url;
 }
