@@ -1,22 +1,24 @@
+require("dotenv").config();
+
 export const config = {
-  "dev": {
-    "username": "",
-    "password": "",
-    "database": "",
-    "host": "",
-    "dialect": "postgres",
-    "aws_region": "us-east-2",
-    "aws_profile": "default",
-    "aws_media_bucket": "udagram-ruttner-dev"
+  dev: {
+    username: process.env.DEV_DATABASE_USERNAME,
+    password: process.env.DEV_DATABASE_PASSWORD,
+    database: process.env.DEV_DATABASE_NAME,
+    host: process.env.DEV_DATABASE_HOST,
+    dialect: process.env.DEV_DATABASE_DIALECT,
+    aws_region: process.env.AWS_REGION,
+    aws_profile: process.env.AWS_PROFILE,
+    aws_media_bucket: process.env.AWS_MEDIA_BUCKET,
   },
-  "jwt": {
-    "secret": " "
+  jwt: {
+    secret: process.env.JWT_SECRET,
   },
-  "prod": {
-    "username": "",
-    "password": "",
-    "database": "udagram_prod",
-    "host": "",
-    "dialect": "postgres"
-  }
-}
+  prod: {
+    username: "",
+    password: "",
+    database: "udagram_prod",
+    host: "",
+    dialect: "postgres",
+  },
+};
