@@ -1,8 +1,15 @@
-import {Table, Column, Model, HasMany, PrimaryKey, CreatedAt, UpdatedAt} from 'sequelize-typescript';
+import {
+  Table,
+  Column,
+  Model,
+  HasMany,
+  PrimaryKey,
+  CreatedAt,
+  UpdatedAt,
+} from "sequelize-typescript";
 
 @Table
-export class User extends Model<User> {
-  
+export class User extends Model {
   @PrimaryKey
   @Column
   public email!: string;
@@ -20,7 +27,7 @@ export class User extends Model<User> {
 
   short() {
     return {
-      email: this.email
-    }
+      email: this.email,
+    };
   }
 }
